feature: Modfieid LLM abstract class
Specifically, the completon abstract function now specifies that it can
accept an optional provision of tools and to ignore it if the model
can't natively handle tools.

The return type is also Any now versus string as we are treating it as
a raw plugin for backends versus direct response parsing. This may be
reverted as I explore more backend work